Bug fixes and network/gps operation enhancements.
Fixed bug where the network would never come up if the device was turned
on in an area without signal.
Changed GPS tick code to stay on until such time as a gps lock is
received or a timeout occurs. Timeout is a simple iteration count.
Added code to set the APN based on the operator name stored on the sim
card.
Added a second thread for sending cell data in the event that the gps
has not locked for a large period of time.
Added device.lua which is aimed at handling operations at a higher layer
than the at_command and at_abs libraries.
Fixed bug in tcp client where the app handles were not being handled
correctly.
Added some example serializer that will come in handy if writing to file
needed at some point.