TCP bug fix and change to default failure count.
Found a bug where the tcp layer would hang if the socket was not cleanly
closed. Would happen if the server was stopped or during times of
network congestion. Fix is to look for any type of close when exiting
the send_data while loop.
Changed the default failure count down to 5 as 10 takes a long time to
reach when reporting once every 30 seconds.