Move http functions out of tcp into own lib. Add http reporter worker.
Move to http reporter model to simplify the main threads. Now the gps
and cell threads will queue data for the http_reporter thread to send
back to the headend.
Move of http specific code out of the tcp_client library. This should
simplify the use of the tcp library and allow the http library to gain
more features.
Removed out command logic in favour of the custom AT commands as they
are much faster.
Started to move all constants into the config library.
Removed MQTT libs for now as not used.