Login on boot and minor loader adjustments.
If session key login is being used then login on boot. Gets around
timing issue where messages can be added before session key information
has been set up.
A couple of changes to try and get the loader to be sync'd for each of
the sio commands. Was getting a situation where the wrong response was
being processed for an sio send. May or may not have fixed this.
Remove some no longer necessary parameters from the headers of the
http lib messages. Encrypted being yes is inferred by the presence
of the key index and imei fields.